Better component names and...  one problem.
Pass async data from parent to child problem is to solve!
