+++ public/src/app/app-main/matrix/row.component.ts
