+++ public/src/app/app-main/table/draws.component.html
@@ -8,7 +8,7 @@
           <th *ngFor="let header of headers">{{ header }}</th>
         </tr>
         <tr>
-          <td *ngFor="let draw of arrayOfDraws">{{ draw }}</td>
+          <td *ngFor="let draw of arrayOfDrawnNumbers">{{ draw }}</td>
         </tr>
       </table>
     </div>
