+++ public/src/app/app-main/wins/output.component.spec.ts
