+++ public/src/app/app-main/wins/output.component.ts
